+++ NEWS.md
@@ -27,6 +27,7 @@ Many bugs have been fixed, and numerous features that were only partially implem
 * Files or folders can now be flagged for exclusion when scraping with the multi-scraper, and for folders it can be set to apply recursively
 * Gamelist sorting is now working as expected and is persistent throughout the application session
 * Overhaul of the game collection functionality including many bug fixes and optimizations
+* Overhaul of the screensaver (the game info overlay now works correctly for instance)
 * Added ability to delete custom collections from the GUI menu
 * Game counting is now done during sorting instead of every time a system is selected. This should make the UI more responsive in case of large game libraries
 * Added a system view counter for favorite games in addition to the total number of games